DRY up countWords implementations
