+++ tests/jest/plugins/tts/utils.test.js
@@ -1,37 +1,5 @@
 import * as utils from '@/src/plugins/tts/utils.js';
 
-describe('approximateWordCount', () => {
-  const { approximateWordCount } = utils;
-
-  test('Empties', () => {
-    expect(approximateWordCount('')).toBe(0);
-    expect(approximateWordCount('    ')).toBe(0);
-  });
-
-  test('Spaceless', () => {
-    expect(approximateWordCount('supercalifragilisticexpialidocious')).toBe(1);
-  });
-
-  test('Basic', () => {
-    expect(approximateWordCount('the quick brown fox jumped over the lazy dog')).toBe(9);
-    expect(approximateWordCount('to be or not to be that is the question')).toBe(10);
-  });
-
-  test('Spacing', () => {
-    expect(approximateWordCount('   the     quick brown   fox jumped over the lazy dog ')).toBe(9);
-  });
-
-  test('Punctuation not separate word (unless separate)', () => {
-    expect(approximateWordCount('the. quick brown. fox. jumped. over the. lazy dog.')).toBe(9);
-    expect(approximateWordCount(' . . . . . ')).toBe(5);
-  });
-
-  test('Realword examples', () => {
-    expect(approximateWordCount("Albert\u2019s cats had a .large blue dish of milk for breakfast.")).toBe(11);
-    expect(approximateWordCount("FARM FOLK O nce upon a time there was a sturdy little boy who lived in Belgium. Every morning after he milked the cows, he gave his cats a large blue dish of milk for breakfast.")).toBe(36);
-  });
-});
-
 describe('toISO6391', () => {
   const { toISO6391 } = utils;
